refactor(analyzers): extract llm helpers to llm-utils
Move extractJSON and findTextRange out of LLMAnalyzer into a sibling module as pure exported functions. Keep the robust JSON parsing and parse-error behavior from main while allowing unit tests to import the helpers directly.
